+++ packages/core/src/structures/Options/ParticleOptions.ts
@@ -10,5 +10,5 @@ export default interface ParticleOptions {
     opacity?: number
     size?: number
     shape?: Shape
-    speed?: Vector
+    velocity?: Vector
 }