+++ packages/core/src/structures/Particle.ts
@@ -1,8 +1,6 @@
 import HEX from "./Colors/HEX"
 import RGBA from "./Colors/RGBA"
 import ParticleSystem from "./ParticleSystem"
-// import FadeOptions from "./Options/FadeOptions"
-// import { FadeInHandler, FadeOutHandler } from "./FadeHandler"
 import ParticleOptions from "./Options/ParticleOptions"
 import ParticleImage from "./ParticleImage"
 import ShapeManager from "./ShapeManager"
@@ -21,31 +19,17 @@ export default class Particle {
     public acceleration: Vector
     public size: number
     public readonly lifeSpan: number
-    public speed: Vector
+    public velocity: Vector
     public color: RGBA | HEX
     public opacity: number
     public shape: Shape
     public get age() { return this._age }
 
-    // Fade Properties
-    // public fadeOut?: FadeOptions --------- Plugin Manager
-    // public fadeIn?: FadeOptions
-    // private fadeOutHandler?: FadeOutHandler
-    // private fadeInHandler?: FadeInHandler
-     
-    public init() {
-        // if(this.fadeOut && (this.fadeOut.opacity != undefined || this.fadeOut.scaleFactor != undefined)) this.fadeOutHandler = new FadeOutHandler(this, this.fadeOut)
-        // if(this.fadeIn && (this.fadeIn.opacity != undefined || this.fadeIn.scaleFactor != undefined)) this.fadeInHandler = new FadeInHandler(this, this.fadeIn)
-
-        // this.opacity = this.fadeIn?.opacity != undefined ? this.fadeIn?.opacity : Math.max(0, Math.min(100, this.opacity))
-        // this.size = this.fadeIn?.scaleFactor != undefined ? this.fadeIn?.scaleFactor : Math.max(0, this.size)
-    }
-
     public update(deltaTime: number) {
-        this.position.x += this.speed.x * deltaTime
-        this.position.y -= this.speed.y * deltaTime
-        this.speed.x += this.acceleration.x * deltaTime
-        this.speed.y += this.acceleration.y * deltaTime
+        this.position.x += this.velocity.x * deltaTime
+        this.position.y -= this.velocity.y * deltaTime
+        this.velocity.x += this.acceleration.x * deltaTime
+        this.velocity.y += this.acceleration.y * deltaTime
 
         for (const meth of this.updateMethods) meth(this, deltaTime)
 
@@ -95,7 +79,7 @@ export default class Particle {
         this.size = options?.size || 0
         this.position = options?.position || { x: 0, y: 0 }
         this.lifeSpan = options?.lifeSpan || 10
-        this.speed = options?.speed || { x: 0, y: 0 }
+        this.velocity = options?.velocity || { x: 0, y: 0 }
         this.color = options?.color || new RGBA(255, 255, 255, 1)
         this.opacity = options?.opacity || 100
         this.shape = options?.shape || 'circle'