+++ packages/core/src/structures/Plugin.ts
@@ -2,7 +2,7 @@ import { ParticleCreateCallback, ParticleUpdateCallback } from "..";
 import ParticleSystem from "./ParticleSystem";
 
 export default interface Plugin {
-    id: string
+    readonly id: string
     setup(system: ParticleSystem): unknown
     onParticleUpdate: ParticleUpdateCallback
     onParticleCreate: ParticleCreateCallback