+++ packages/core/types/structures/Colors/HEX.d.ts
@@ -1,6 +1,6 @@
 import RGBA from "./RGBA";
 export default class HEX {
-    hex: string;
+    private hex;
     constructor(hex: string);
     toRGB(): RGBA;
     toString(): string;