+++ packages/particulab/src/index.ts
@@ -1,2 +1,3 @@
 export * from '@particulab/core'
 export * from '@particulab/fade'
+export * from '@particulab/reactivity'