+++ packages/reactivity/src/ParticleReactivityOptions.ts
@@ -1,0 +1,5 @@
+import { Interval } from "@particulab/core";
+
+export default interface ParticleReactivityOptions {
+    mass: Interval
+}