+++ packages/reactivity/src/ReactivityOptions.ts
@@ -1,0 +1,14 @@
+import { Interval } from "@particulab/core";
+import { ReactionMode } from ".";
+import ParticleReactivityOptions from "./ParticleReactivityOptions";
+
+
+export default interface ReactivityOptions {
+    mode: ReactionMode
+    radius: number
+    /**
+     * @param strength Force strength when `mode` is `attract` or `repel`.
+     */
+    strength?: number
+    particleConfig?: ParticleReactivityOptions
+}