+++ packages/reactivity/types/ParticleReactivityOptions.d.ts
@@ -1,0 +1,4 @@
+import { Interval } from "@particulab/core";
+export default interface ParticleReactivityOptions {
+    mass: Interval;
+}